move uppy_dashboard to webpacker
Don't need to hide functions in a closure, webpacker does that for all included source files
